fix: adding a new verification to the board delete

+++ src/domain/features/delete-board.ts
@@ -19,6 +19,11 @@ export const setupDeleteBoard: Setup =
   ({ repository }) =>
   async ({ board, id }) => {
     try {
+      if(board.deleted === 0){
+        throw new Error("Board must be marked as deleted before deletion", {
+          cause: "board-not-deleted",
+        });
+      }
       await repository.deleteBoard(board, id);
       return {
         statusCode: 200,
